feat: move C decls to cdecl
